# src/storage/api-docs/api-docs.json
{
    "openapi": "3.0.0",
    "info": {
        "title": "Documentation API Library",
        "version": "1.0"
    },
    "paths": {
...
        "/api/books": {
            "get": {
                "tags": [
                    "Books"
                ],
                "summary": "Get a paginated list of books",
                "description": "Fetch a paginated list of books with details like total data, per page, current page, last page, and next page URL",
                "operationId": "getBooks",
                "parameters": [
                    {
                        "name": "page",
                        "in": "query",
                        "description": "The page number to retrieve",
                        "required": false,
                        "schema": {
                            "type": "integer"
                        }
                    }
                ],
                "responses": {
                    "200": {
                        "description": "Books retrieved successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "integer",
                                            "example": 200
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "index"
                                        },
                                        "data": {
                                            "properties": {
                                                "total_data": {
                                                    "type": "integer",
                                                    "example": 500
                                                },
                                                "per_page": {
                                                    "type": "integer",
                                                    "example": 20
                                                },
                                                "current_page": {
                                                    "type": "integer",
                                                    "example": 1
                                                },
                                                "last_page": {
                                                    "type": "integer",
                                                    "example": 25
                                                },
                                                "next_page_url": {
                                                    "type": "string",
                                                    "example": "http://localhost:8004/api/books?page=2"
                                                },
                                                "result": {
                                                    "type": "array",
                                                    "items": {
                                                        "properties": {
                                                            "id": {
                                                                "type": "integer",
                                                                "example": 1
                                                            },
                                                            "title": {
                                                                "type": "string",
                                                                "example": "Prof."
                                                            },
                                                            "description": {
                                                                "type": "string",
                                                                "example": "Alice. 'Of course twinkling begins with an air of great dismay..."
                                                            },
                                                            "publish_date": {
                                                                "type": "string",
                                                                "format": "date",
                                                                "example": "2002-12-22"
                                                            },
                                                            "author_id": {
                                                                "type": "integer",
                                                                "example": 33
                                                            }
                                                        },
                                                        "type": "object"
                                                    }
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            },
            "post": {
                "tags": [
                    "Books"
                ],
                "summary": "Create a new book",
                "description": "Create a new book entry in the database",
                "operationId": "createBook",
                "requestBody": {
                    "required": true,
                    "content": {
                        "application/json": {
                            "schema": {
                                "properties": {
                                    "title": {
                                        "type": "string",
                                        "example": "Filosofi Teras"
                                    },
                                    "description": {
                                        "type": "string",
                                        "example": "Kita memiliki kebiasaan membesar-besarkan kesedihan. ..."
                                    },
                                    "publish_date": {
                                        "type": "string",
                                        "format": "date",
                                        "example": "2018-01-01"
                                    },
                                    "author_id": {
                                        "type": "integer",
                                        "example": 3
                                    }
                                },
                                "type": "object"
                            }
                        }
                    }
                },
                "responses": {
                    "201": {
                        "description": "Book created successfully",
                        "content": {
                            "application/json": {
                                "schema": {
                                    "properties": {
                                        "status": {
                                            "type": "boolean",
                                            "example": true
                                        },
                                        "code": {
                                            "type": "string",
                                            "example": "201"
                                        },
                                        "text": {
                                            "type": "string",
                                            "example": "success"
                                        },
                                        "method": {
                                            "type": "string",
                                            "example": "store"
                                        },
                                        "data": {
                                            "properties": {
                                                "title": {
                                                    "type": "string",
                                                    "example": "Filosofi Teras"
                                                },
                                                "description": {
                                                    "type": "string",
                                                    "example": "Kita memiliki kebiasaan membesar-besarkan kesedihan. ..."
                                                },
                                                "publish_date": {
                                                    "type": "string",
                                                    "format": "date",
                                                    "example": "2018-01-01"
                                                },
                                                "author_id": {
                                                    "type": "integer",
                                                    "example": 3
                                                },
                                                "id": {
                                                    "type": "integer",
                                                    "example": 502
                                                }
                                            },
                                            "type": "object"
                                        }
                                    },
                                    "type": "object"
                                }
                            }
                        }
                    }
                }
            }
        },
...
    },
...
}